requirements.txt
python-telegram-bot==20.7
langchain==0.1.0
openai==1.12.0
python-dotenv==1.0.0
biopython==1.83
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
tqdm==4.66.1 